Internals

RegistryCI.test — Function

test(path)

Run various checks on the registry located at path. Checks for example that all files are parsable and understandable by Pkg and consistency between Registry.toml and each Package.toml.

If your registry has packages that have dependencies that are registered in other registries elsewhere, then you may provide the github urls for those registries using the registry_deps parameter.

RegistryCI.AutoMerge.get_all_non_jll_package_names — Method

get_all_non_jll_package_names(registry_dir::AbstractString) -> Vector{String}

Given a path to the directory holding a registry, returns the names of all the non-JLL packages defined in that registry, along with the names of Julia's standard libraries.

RegistryCI.AutoMerge.load_files_from_url_and_tree_hash — Method

load_files_from_url_and_tree_hash(f, destination::String, url::String, tree_hash::String) -> Bool

Attempts to clone a git repo from url into a temporary directory, runs f(dir) on that directory, then extract the files and folders from a given tree_hash, placing them in destination.

Returns a boolean indicating if the cloning succeeded.

RegistryCI.AutoMerge.parse_registry_pkg_info — Function

parse_registry_pkg_info(registry_path, pkg, version=nothing) -> @NamedTuple{uuid::String, repo::String, subdir::String, tree_hash::Union{Nothing, String}}

Searches the registry located at registry_path for a package with name pkg. Upon finding it, it parses the associated Package.toml file and returns the UUID and repository URI, and subdir.

If version is supplied, then the associated tree_hash will be returned. Otherwise, tree_hash will be nothing.

RegistryCI.AutoMerge.run — Function

run(env, cicfg::CIService; kwargs...)