Public API

RegistryCI.test — Function

test(path)

Run various checks on the registry located at path. Checks for example that all files are parsable and understandable by Pkg and consistency between Registry.toml and each Package.toml.

If your registry has packages that have dependencies that are registered in other registries elsewhere, then you may provide the github urls for those registries using the registry_deps parameter.

RegistryCI.AutoMerge.run — Function

run(env, cicfg::CIService; kwargs...)