Public API

RegistryCI

RegistryCI.test — Function

test(path)

Run various checks on the registry located at path. Checks for example that all files are parsable and understandable by Pkg and consistency between Registry.toml and each Package.toml.

If your registry has packages that have dependencies that are registered in other registries elsewhere, then you may provide the github urls for those registries using the registry_deps parameter.

AutoMerge

AutoMerge.check_pr — Function

check_pr(; registry, authorized_authors, ..., env=ENV, cicfg=auto_detect_ci_service(; env=env))

Check a pull request for registration validity. This entrypoint runs untrusted code and does not require merge permissions.

Required Arguments

registry::String: the registry name you want to run AutoMerge on.

authorized_authors::Vector{String}: list of who can submit registration, e.g String["JuliaRegistrator"].

authorized_authors_special_jll_exceptions::Vector{String}: a list of users who can submit JLL packages.

new_package_waiting_minutes::Dates.Minute: new package waiting period in minutes.

new_jll_package_waiting_minutes::Dates.Minute: new JLL package waiting period in minutes.

new_version_waiting_minutes::Dates.Minute: new package version waiting period in minutes.

new_jll_version_waiting_minutes::Dates.Minute: new JLL package version waiting period in minutes.

Optional Arguments

master_branch::String = "master": name of master_branch.

error_exit_if_automerge_not_applicable::Bool = false: if false, AutoMerge will not error on PRs made by non-AutoMerge-authorized users

api_url::String = "https://api.github.com": the registry host API URL.

read_only::Bool = false: run in read only mode.

master_branch_is_default_branch::Bool = true: if master_branch is the default branch.

suggest_onepointzero::Bool = true: should the AutoMerge comment include a suggestion to tag a 1.0 release.

point_to_slack::Bool = false: should the AutoMerge comment recommend sending a message to the #pkg-registration Julia-Slack channel.

registry_deps::Vector{<:AbstractString} = String[]: list of registry dependencies.

check_license::Bool = false: check package has a valid license.

check_breaking_explanation::Bool = false: Check whether the PR has breaking change explanation.

public_registries::Vector{<:AbstractString} = String[]: registries to check for UUID collisions.

environment_variables_to_pass::Vector{<:AbstractString} = String[]: Environment variables to pass to subprocess.

commit_status_token_name::String = "AUTOMERGE_GITHUB_TOKEN": Name of environment variable containing GitHub token.

env: an AbstractDictionary used to read environmental variables from.

cicfg: Configuration struct describing the continuous integration (CI) environment.

AutoMerge.merge_prs — Function

merge_prs(; registry, authorized_authors, ..., env=ENV, cicfg=auto_detect_ci_service(; env=env))

Merge approved pull requests. This entrypoint requires merge permissions and does not run untrusted code.

Required Arguments

registry::String: the registry name you want to run AutoMerge on.

authorized_authors::Vector{String}: list of who can submit registration, e.g String["JuliaRegistrator"].

authorized_authors_special_jll_exceptions::Vector{String}: a list of users who can submit JLL packages.

new_package_waiting_minutes::Dates.Minute: new package waiting period in minutes.

new_jll_package_waiting_minutes::Dates.Minute: new JLL package waiting period in minutes.

new_version_waiting_minutes::Dates.Minute: new package version waiting period in minutes.

new_jll_version_waiting_minutes::Dates.Minute: new JLL package version waiting period in minutes.

Optional Arguments

master_branch::String = "master": name of master_branch.

error_exit_if_automerge_not_applicable::Bool = false: if false, AutoMerge will not error on PRs made by non-AutoMerge-authorized users

api_url::String = "https://api.github.com": the registry host API URL.

read_only::Bool = false: run in read only mode.

merge_new_packages::Bool = true: should AutoMerge merge registration PRs for new packages

merge_new_versions::Bool = true: should AutoMerge merge registration PRs for new versions of packages

additional_statuses::AbstractVector{<:AbstractString} = String[]: list of additional commit statuses that must pass

additional_check_runs::AbstractVector{<:AbstractString} = String[]: list of additional check runs that must pass

merge_token_name::String = "AUTOMERGE_MERGE_TOKEN": Name of environment variable containing GitHub token for PR merging.

env: an AbstractDictionary used to read environmental variables from.

cicfg: Configuration struct describing the continuous integration (CI) environment.

AutoMerge.check_pr — Method

check_pr(; registry, authorized_authors, ..., env=ENV, cicfg=auto_detect_ci_service(; env=env))

Check a pull request for registration validity. This entrypoint runs untrusted code and does not require merge permissions.

Required Arguments

registry::String: the registry name you want to run AutoMerge on.

authorized_authors::Vector{String}: list of who can submit registration, e.g String["JuliaRegistrator"].

authorized_authors_special_jll_exceptions::Vector{String}: a list of users who can submit JLL packages.

new_package_waiting_minutes::Dates.Minute: new package waiting period in minutes.

new_jll_package_waiting_minutes::Dates.Minute: new JLL package waiting period in minutes.

new_version_waiting_minutes::Dates.Minute: new package version waiting period in minutes.

new_jll_version_waiting_minutes::Dates.Minute: new JLL package version waiting period in minutes.

Optional Arguments

master_branch::String = "master": name of master_branch.

error_exit_if_automerge_not_applicable::Bool = false: if false, AutoMerge will not error on PRs made by non-AutoMerge-authorized users

api_url::String = "https://api.github.com": the registry host API URL.

read_only::Bool = false: run in read only mode.

master_branch_is_default_branch::Bool = true: if master_branch is the default branch.

suggest_onepointzero::Bool = true: should the AutoMerge comment include a suggestion to tag a 1.0 release.

point_to_slack::Bool = false: should the AutoMerge comment recommend sending a message to the #pkg-registration Julia-Slack channel.

registry_deps::Vector{<:AbstractString} = String[]: list of registry dependencies.

check_license::Bool = false: check package has a valid license.

check_breaking_explanation::Bool = false: Check whether the PR has breaking change explanation.

public_registries::Vector{<:AbstractString} = String[]: registries to check for UUID collisions.

environment_variables_to_pass::Vector{<:AbstractString} = String[]: Environment variables to pass to subprocess.

commit_status_token_name::String = "AUTOMERGE_GITHUB_TOKEN": Name of environment variable containing GitHub token.

env: an AbstractDictionary used to read environmental variables from.

cicfg: Configuration struct describing the continuous integration (CI) environment.

AutoMerge.merge_prs — Method

merge_prs(; registry, authorized_authors, ..., env=ENV, cicfg=auto_detect_ci_service(; env=env))

Merge approved pull requests. This entrypoint requires merge permissions and does not run untrusted code.

Required Arguments

registry::String: the registry name you want to run AutoMerge on.

authorized_authors::Vector{String}: list of who can submit registration, e.g String["JuliaRegistrator"].

authorized_authors_special_jll_exceptions::Vector{String}: a list of users who can submit JLL packages.

new_package_waiting_minutes::Dates.Minute: new package waiting period in minutes.

new_jll_package_waiting_minutes::Dates.Minute: new JLL package waiting period in minutes.

new_version_waiting_minutes::Dates.Minute: new package version waiting period in minutes.

new_jll_version_waiting_minutes::Dates.Minute: new JLL package version waiting period in minutes.

Optional Arguments

master_branch::String = "master": name of master_branch.

error_exit_if_automerge_not_applicable::Bool = false: if false, AutoMerge will not error on PRs made by non-AutoMerge-authorized users

api_url::String = "https://api.github.com": the registry host API URL.

read_only::Bool = false: run in read only mode.

merge_new_packages::Bool = true: should AutoMerge merge registration PRs for new packages

merge_new_versions::Bool = true: should AutoMerge merge registration PRs for new versions of packages

additional_statuses::AbstractVector{<:AbstractString} = String[]: list of additional commit statuses that must pass

additional_check_runs::AbstractVector{<:AbstractString} = String[]: list of additional check runs that must pass

merge_token_name::String = "AUTOMERGE_MERGE_TOKEN": Name of environment variable containing GitHub token for PR merging.

env: an AbstractDictionary used to read environmental variables from.

cicfg: Configuration struct describing the continuous integration (CI) environment.