Public API

RegistryCI

RegistryCI.test — Function

test(path)

Run various checks on the registry located at path. Checks for example that all files are parsable and understandable by Pkg and consistency between Registry.toml and each Package.toml.

If your registry has packages that have dependencies that are registered in other registries elsewhere, then you may provide the github urls for those registries using the registry_deps parameter.

AutoMerge

Main Entrypoints

AutoMerge.check_pr — Function

check_pr(registry_config::RegistryConfiguration, pr_config::CheckPRConfiguration, env=ENV, cicfg::CIService=auto_detect_ci_service(; env=env))

Check a pull request for registration validity. This entrypoint runs untrusted code and does not require merge permissions.

Arguments

registry_config: RegistryConfiguration struct containing shared registry settings

pr_config: CheckPRConfiguration struct containing PR validation specific settings

env: an AbstractDictionary used to read environmental variables from. Defaults to ENV but a plain Dict can be passed to mimic an alternate environment.

cicfg: Configuration struct describing the continuous integration (CI) environment in which AutoMerge is being run.

Example

Here is an example of how General registry is configured:

using AutoMerge
(; registry_config, check_pr_config) = AutoMerge.general_registry_config()
AutoMerge.check_pr(registry_config, check_pr_config)

To configure a custom registry, save a .toml configuration file somewhere. This can be based on the one General uses, which you can obtain by

config = AutoMerge.general_registry_config()
AutoMerge.write_config("AutoMerge.toml", config)

and then modify to suit your needs. This can then be used via:

using AutoMerge
(; registry_config, check_pr_config) = AutoMerge.read_config("path/to/AutoMerge.toml")
AutoMerge.check_pr(registry_config, check_pr_config)

AutoMerge.merge_prs — Function

merge_prs(registry_config::RegistryConfiguration, merge_config::MergePRsConfiguration, env=ENV, cicfg::CIService=auto_detect_ci_service(; env=env))

Merge approved pull requests. This entrypoint requires merge permissions and does not run untrusted code.

Arguments

registry_config: RegistryConfiguration struct containing shared registry settings

merge_config: MergePRsConfiguration struct containing merge specific settings

env: an AbstractDictionary used to read environmental variables from. Defaults to ENV but a plain Dict can be passed to mimic an alternate environment.

cicfg: Configuration struct describing the continuous integration (CI) environment in which AutoMerge is being run.

Example

Here is an example of how General registry is configured:

using AutoMerge

(; registry_config, merge_prs_config) = AutoMerge.general_registry_config()
AutoMerge.merge_prs(registry_config, merge_prs_config)

Configuration Types

For configuration information, see the Configuration page.